Fixed No Fuel
if you ran out of fuel all movement inputs would stop working so the player would just endlessly move in one direction without being able to influence it, this fix automatically turns off fuel usage when you run out of fuel so you can keep moving.

Also moved the the rotation from Update to FixedUpdate to avoid tying any of the physics to the fps which is a good practice.